requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.5
django-templated-mail==1.1.1
djangorestframework==3.10.3
djoser==2.0.3
idna==2.8
psycopg2==2.8.3
pytz==2019.2
requests==2.22.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.3
wikipedia==1.4.0
